+++ src/lib.rs
@@ -70,6 +70,7 @@ pub use crate::algorithms::properties::betweenness_weighted::betweenness_weighte
 pub use crate::algorithms::properties::closeness::closeness;
 pub use crate::algorithms::properties::closeness_weighted::closeness_weighted;
 pub use crate::algorithms::properties::edge_betweenness::edge_betweenness;
+pub use crate::algorithms::properties::edge_betweenness_weighted::edge_betweenness_weighted;
 pub use crate::algorithms::properties::eigenvector::eigenvector_centrality;
 pub use crate::algorithms::properties::girth::girth;
 pub use crate::algorithms::properties::harmonic::harmonic_centrality;